reduce test noise

+++ air/lib/air/service/shadow_db/server.ex
@@ -15,23 +15,25 @@ defmodule Air.Service.ShadowDb.Server do
   @doc "Drops the given shadow database."
   @spec drop_database(String.t()) :: :ok
   def drop_database(data_source_name) do
-    {:ok, conn} = Postgrex.start_link(hostname: "127.0.0.1", username: "postgres", database: "postgres")
+    if Application.get_env(:air, :shadow_db?, true) do
+      {:ok, conn} = Postgrex.start_link(hostname: "127.0.0.1", username: "postgres", database: "postgres")
 
-    try do
-      # force close all existing connections to the database
-      Postgrex.query(
-        conn,
-        """
-        SELECT pg_terminate_backend(pg_stat_activity.pid)
-        FROM pg_stat_activity
-        WHERE pg_stat_activity.datname = $1 AND pid <> pg_backend_pid();
-        """,
-        [db_name(data_source_name)]
-      )
+      try do
+        # force close all existing connections to the database
+        Postgrex.query(
+          conn,
+          """
+          SELECT pg_terminate_backend(pg_stat_activity.pid)
+          FROM pg_stat_activity
+          WHERE pg_stat_activity.datname = $1 AND pid <> pg_backend_pid();
+          """,
+          [db_name(data_source_name)]
+        )
 
-      Postgrex.query(conn, ~s/DROP DATABASE IF EXISTS "#{sanitize_name(db_name(data_source_name))}"/, [])
-    after
-      close_conn(conn)
+        Postgrex.query(conn, ~s/DROP DATABASE IF EXISTS "#{sanitize_name(db_name(data_source_name))}"/, [])
+      after
+        close_conn(conn)
+      end
     end
 
     :ok
@@ -50,16 +52,18 @@ defmodule Air.Service.ShadowDb.Server do
 
   @impl GenServer
   def handle_cast(:update_definition, state) do
-    case data_source_tables(state.data_source_name) do
-      [] ->
-        drop_database(state.data_source_name)
-        {:stop, :normal, state}
+    if Application.get_env(:air, :shadow_db?, true) do
+      case data_source_tables(state.data_source_name) do
+        [] ->
+          drop_database(state.data_source_name)
+          {:stop, :normal, state}
 
-      tables ->
-        if Application.get_env(:air, :shadow_db?, true) and state.tables != tables,
-          do: update_shadow_db(state, tables)
-
-        {:noreply, %{state | tables: tables}}
+        tables ->
+          if state.tables != tables, do: update_shadow_db(state, tables)
+          {:noreply, %{state | tables: tables}}
+      end
+    else
+      {:noreply, state}
     end
   end
